requirements.txt
# Core
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

# Detection
ultralytics>=8.2.0       # YOLOv8 / YOLO11 support
opencv-python>=4.10.0
numpy>=1.26.0

# OCR & Image Processing
Pillow>=10.2.0
scikit-image>=0.22.0

# Data Science
pandas>=2.2.0
scikit-learn>=1.4.0

# Visualization & Debug
matplotlib>=3.8.0
seaborn>=0.13.0

# Logging & Progress
tqdm>=4.66.0
loguru>=0.7.0

# Database (for license_plates.db)
sqlalchemy>=2.0.0

# Testing / Debug
pytest>=8.0.0

# Optional (for SORT tracker, sometimes needed)
filterpy>=1.4.5
